Client feeback
After consulting my client/stakeholder, we have come to the conclusion that we wanted to focus on information in regards to event happening this year the sport sailing is not currently happen to our knowledge at the school. So for the this years iteration i have chosen to keep it out of the website.