Change of layout and CSS
Removed all my Header, Nav, Footer components from all my pages and moved it into my Layout css. More of an upgrade so that i know that my components are all on one page and limits the amount of changes i make it overall.

Then for the CSS i have moved the text margins so that you can see the bullet points and so that the text does not over the page.

Basically trialed and tested how the content displays on the pages