Updates
Updated a few missing images and made sure that all of the TIC were at the top of the page with their appropriate image. I made this choice as I felt that  the information was quite important and believed it need to stand out. After having it being displayed in a similar format as the current sports portal. I felt it was the better choice to have it displayed that way.

Trying to test and see what other ways I could display the Navigation bar