Design/ Content Change
Basically focused and tested different ways to make sure that the information and images were balanced at eye level. Found a way to make it as consistent as possible.

Change the size of the staff names and information as making it smaller just looked more appealing and felt like the better close when seeing and consider how it flows with the rest of the information on the page.

Removed some content that I forgotten to do previously.

basically made all of the changes after testing  multiple ways that the information could be displayed.