Update and Applied feedback
Applied feedback and had users test it to see if it works.

Also tested  to see if the website works on on multiple browsers. The display was consistent throughout and work as expected on chrome, safari and edge.

Lastly there was a test to see how it would be displayed on mobile devices and that was  successful.