Implications
Added images to the website and changed the size of them to be look suitable and be consistent throughout the entire website. This was applied feedback from a user. Which they expressed they like the way images are being displayed now as its easy on the eye. The website was tested and used by users and was done successfully.

While on applying the feedback to the website, I have began considering the implications throughout the website.

When making use of images I made sure that I was inclusive of the people that I was displaying. The way I sourced my images was from public platforms which student consent was already granted. When using staff and code captain images, I emailed for consent prior to using them. This covers my social/ cultural responsibility along with the legal obligations im meant to follow.

Throughout my internal I had to get granted access from my stakeholders to make use of most of my content.

In regards to making my website aesthetic I tested and tried different ways to include the school colours while trying to maintain it formality.

Through the testing of the users I have seen that the website functions as expected. This is through multiple browsers. With all the pages being displayed correctly and respond and works the way its supposed to.

The layout that I have chosen is something that can be easily read by all users. The sizes of my font and the way link are displayed are common ways tat the user would have already been exposed to.

The website is accessible for many, it has been checked through sim daltonism. On mobile devices, on a variety of browsers.