Implications- indepth
Acessbility

There were many steps I had to go through to ensure that my web app is accessible. As its important that my website is usable and accessible by everyone including this who may have disabilities. The images that are displayed on the page all come with an alt text. As in some cases screen readers are unable to read images so an alt text helps describe the images to the users when needed.

When choosing my colours on my website I had to ensure that the colours I was picking, contrasted with one another to allow for users to easily see and distinguish the words in the page. With this my choice of fonts were quite bubbly and space which Mr Jones , Mr Muholland , Kate davies and Tercy Chitima felt were easily on the eye. Which was the goal. Accomodating suers who may have low vision.

My choice of colours for my font were black and a navy blue because they have the highest readability. Because my website is open and available to many I have to ensure that its readability is high without worrying whether it would be accessible to everyone.

Another way  my website is made accessible is through the use of semantic html. These element helps define the meaning and purpose of the content that has been added tip the web page , beaus elf this screen readers or other assertive technologies know the structure of the page and can now provide a better experience for those who rely on assistance to use technology.

My choice to use headings and subheading helps in organising the content of the page and provides the users with a clear structure which is also favoured by assistive technologies. These headings  create a hierarchy of information. So that its easier on people who struggle to process large blocks of information to understand.

Usability
The URLs are extremely detailed and offer enough detail to the user.
1. I made my urls looks like buttons  because if I just displayed the link on the page the users would not know what the link was intended to. Displaying it as a button made it easier fro me to describe the element.
2. The pages in the name those URLS are quite descriptive and straight forward. Users who were using screen reader would know what the pages would be covering because of their names

Speaking onthis I have tested what my website looks like through the different settings on sim daltonism to ensure that my website and the choice of colours were accommodative of people with disabilities. Making it  usable for many.

Functionality
I have tested my website with 5 real users to try and find out if there were any issues that they may have come across. Their feedback allowed for me to make necessary improvements to make sure that  my website can at least be seen as usable or practical by most people.

Due to the increase amount of technology and devices that people would and couple potentially access my website through. Its important that I tested what it would display on multiple devices. To ensure that everyone would go through the same experiences and accommodate those who may struggle with technology. I believe that my design is responsive and can adapt to the proved device. Appearing as consistent and accessible to the users.

When presenting my website with users to test it out , I watched to see if the website auctioned the way that I intended it to work. Aswell as ensured that users were aware of the way it was meant to function to avoid any confusion in the future.

I have had to test my website on Multiple browsers to ensure that one it was displayed the same on all and second that it worked because if not code would have been displayed on the suers page.

Future proofing
Accessibility is not something that can be done and never improved again, with the help of my readme id be able to develop ways to make my website accessible for users going forward.